pyproject.toml
[tool.poetry]
name = "code"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
ipykernel = "^6.20.1"
plotly = "^5.11.0"
matplotlib = "^3.6.2"
nsetools = "^1.0.11"
backtesting = "^0.3.3"
nsepy = "^0.8"
openpyxl = "^3.1.0"
bokeh = "2.4.3"
streamlit = "^1.17.0"
pandas-ta = "^0.3.14b0"
black = "^23.1.0"
pandas-market-calendars = "^4.1.4"
humanize = "^4.6.0"
schedule = "^1.1.0"
coloredlogs = "^15.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
